+++ www/admin/templates/view_contest.tpl
@@ -18,6 +18,13 @@
 <td>{$images[img].image.display_name}</td>
 <td>{$images[img].image.vote_count}</td>
 <td>
+{if $images[img].image.must_win eq 0}
+    Не задано
+{elseif $images[img].image.must_win eq 1}
+    Да
+{/if}
+</td>
+<td>
 {section name=vote loop=$images[img].votes}
 {$images[img].votes[vote].display_name} ({$images[img].votes[vote].user_id} : <a href="http://whatismyipaddress.com/ip/{$images[img].votes[vote].from}" target="_blank">{$images[img].votes[vote].from}</a>)<br>
 {/section}
@@ -31,7 +38,7 @@
 </td>
 <td>
 <a href="http://{$smarty.server.SERVER_NAME}{$SCRIPT_NAME}?contest={$images[img].image.contest_id}&id={$images[img].image.id}&type=image&action=edit">Редактировать</a> |
-<a href="http://{$smarty.server.SERVER_NAME}{$SCRIPT_NAME}?contest={$images[img].image.contest_id}&id={$images[img].image.id}&type=image&action=delete">Удалить</a>
+<a href="http://{$smarty.server.SERVER_NAME}{$SCRIPT_NAME}?contest={$images[img].image.contest_id}&id={$images[img].image.id}&type=image&action=delete" onclick="return confirm('Удалить?');">Удалить</a>
 </td>
 </tr>
 {/section}
